Update OrganisationFetcher to handle recreations
The API can return an organisation with a slug that matches one we
already know about, however the content_id may differ from the one we
hold locally. This can happen if within the API the organisation has
been destroyed and then recreated with the same slug. In this instance
the organisation is assigned a new content_id by the API and causes
a conflict with our internal representation when we try to sync.