Revoke ApiUser tokens with POST request
Revoking / re-generating tokens is a privileged action within Signon,
only available to superadmin users.

It's better to use a POST (or at least, non-GET) for this as GET
requests with no additional headers can be initiated from a different
origin, see Cross-Origin Resource Sharing for details [0].

This change updates the method to be a POST and uses a standard rails
`form_tag` which provides Cross-Site Request Forgery tokens as standard.

[0] - https://en.wikipedia.org/wiki/Cross-origin_resource_sharing