Let Devise determine redirect path after reset
This was redirecting the user to /, which because the user was
unauthorised was redirecting the user to the login page. In the interim
redirect, the flash message confirming passphrase reset success was
overwritten by a “you must login” message. Devise will automatically
redirect to sign in, meaning the success flash message works again.