Clean expired oauth_access_tokens records
This repurposes (and makes more efficient) the rake task that clears
out the OAuth Access Grants table of expired rows to also be used for
OAuth Access Tokens. The tokens table is around 5 million rows in
production and weighs about 5GB. We don’t need expired records,
certainly not ones from 2012. This slows down nightly replication from
production to staging, and from integration to developer laptops.

The deletion process has changed to fetch deletable primary keys in
batches of 1000, then delete them in those batches. This prevents the
table locking for prolonged periods.

One rake task now cleans both tables, and the old rake task points to
the new one. The puppet code that defines the cronjob for this should
be updated, but this redirection makes the change backwards compatible.