Explicitly set SSOPushCredential user in test
Calling `SSOPushCredential.credentials` results in adding a set of
permissions to the `User` object assigned to the class’ `user`
accessor. In tests, this may not be the same object as previously set,
and might refer to a different record in the database.

This caused the `PermissionsUpdater` tests to fail as it explicitly
tests for the permissions on a `User` object, which happened to have
permissions added to it by `SSOPushCredential`.