Extend grace period for newly unsuspended users
Before this functionality was added, the following would happen:

1. User reaches inactivity time and gets automatically suspended at 5am
2. User is unsuspended but doesn't yet log in
3. User gets re-suspended at 5am the next day

This grace period was originally set at 3 days to combat this. However
it causes issues where people are off a day either side of a weekend
and is still generating repeat tickets. We’re going to try 7 days to
see if it reduces the number of support requests we receive.