Add centralized location for default permissions for users
We add `Signon.default_permissions_for_all_users` to hold a list of the
default permissions that should be added to all users.  We then make the
three code paths that create new users use this mechanism.  The defaults
are added in an intializer which is hard coded for now; a deploy and
restart is required to change the defaults.  Future work could be done
to add a UI on this.

Note that the intializer actually does its work in an `after_initialize`
callback so that we know all our dependencies (doorkeeper and our
enhancements to it) have been loaded and are ready to work.  We also
don't add the default permission in test (because it would run at
a time before the test db is neccessarily prepared); if you need it be
sure to set it up in your tests specifically.