Add hint about locked accounts to signin failure warning
For security reasons we don't tell the person entering credentials that an
account has been locked when accessing signon. That would reveal that the
details entered are for an actual account.

We can however suggest to the user that after X attempts the account _may_
now be locked, so that's what we do here.

There are 3 possible error messages you'd get for failing a login: locked,
invalid, and not_found_in_database.  We've changed all of these to the
same generic "failure, maybe locked?" message.  We are running devise in
paranoid mode so it will only use the invalid message, but we change all
of them just in case.