Introduce specs for Policy scopes
We also make a tiny change to the implementation to return scope.none
which is AR convention for something that will return no results, but
still act like an AR object for the purposes of chaining other methods
and API calls.  This means asking for a scope for a user that has no
permissions to see anything won't get `nil` or an empty array and risk
`NoMethodError`s.