Stop org admins from viewing organisations
Org admins used to be able to see a list of all orgs under their own orgs
subtree, but since we removed the ability for them to edit those users or
add those orgs to a user it doesn't make sense for them to see this list.
Not least because this list is somewhat useless on it's own.