Introduce UserPermissionManageableApplicationPolicy
This wraps up the code in UserControllerMethods#visible_applications for
web users as a pundit policy to make it easier to reuse.  The policy
doesn't have any permission methods on it, it just defines a scope.
The name is slightly unweildy, but I struggled to come up with something
shorter that didn't sound too generic.  It has been chosen to give a hint
that it refers to the `manage_supported_permissions?` permission on the
`UserPolicy`.  The scope gives us the list of applications that the
current user is allowed to edit permissions of.  For example, a
superadmin can edit permissions for all applications, and so can an admin
user.  Org admins however can only edit the permissions of applications
that allow delegation of permissions and that the org admin has access
to themselves.  Normal users can't edit permissions for any apps.