Limit the permissions that org admins can edit when updating users
Org admins are presented a limited set of permissions to manage when they
edit a user.  Prior to this change submitting that form would set the
users full set of permissions to the limited set available to the org
admin.  Meaning that if a superadmin granted a user permission on an app
that the org admin was not allowed to edit, this permission would be lost
when the org admin edits the user.

This change makes sure that when we update a user we only remove and add
permissions that the acting user is allowed to manage.  We introduce a new
SupportedPermissionPolicy with a scope to tell us which permissions a given
user is allowed to manage (making use of the previously added
UserPermissionManageableApplicationPolicy scope for org admins to get
the correct set of applications).

It's not 100% clear to me that I have decent test coverage here as the
integration test by design has to stop at the UI layer when checking what
permissions a user can / can't grant to another user.

It's also unclear what the expected effect of the grantable_from_ui flag
on a permission is.  It's used to limit the set of permissions available
in the view layer, but so far not to limit what permissions the acting
user is allowed to manage on a user object.  Perhaps it should be rolled
into the policies too?

fixup! Limit the permissions that org admins can edit when updating users