Require signin permission to complete oauth authorisation
When an application makes an authorisation request on behalf of a user
that user must have the `signin` permission granted to them for the
request to be completed successfully.  If they don't the user will be
redirected to a "sorry, you don't have permission" page on signon. This
new page is based on the error page from the gds-sso gem that serves the
same purpose.

We have to configure doorkeeper to use a custom authorizations controller
to achieve this.  The `resource_owner_authenticator` block where we check
if the user needs 2 step verification or to change their password or other
user based reasons to interrupt the login flow would appear to be the
pefect place to do the `signin` permission check, unfortunately however
this block is run before doorkeeper has extracted the application from the
requset so we don't have access to it.  The `skip_authorization` block is
yielded the user and application (inside the client), but this is only run
for the auth/new path and is not expected to have side-effects like
rendering or redirecting.  Instead we use our own controller and add a
branch for permission check.

The test provided for the custom controller is a port of the test in
doorkeeper for the default authorizations controller, translated into
minitest and given some extra config to run in our suite insetead of
the doorkeeper one.  We've also added test cases to cover our permissions
functionality.

We'll need to be careful to make sure that when we update doorkeeper we
port any changes to the default authorisations controller or test to our
versions.  If a hook becomes available it would be a good idea to use
that if possible.