Add signin permission requirement to doorkeeper_authorize!
For endpoints made available to oauth clients we make sure that the user
that owns the access token used on the request has signin permission on
the application making the request on their behalf.  To achieve this we
customise the `doorkeeper_authorize!` method that doorkeeper provides to
protect your controllers.  Our version calls super to let the default
implementation take care of parsing tokens and doing the real work of
authenticating the request.  Once this is complete we check that the
user has signin permission on the applicaiton (we obtain the user and
application from the access token).  If not we redirect to the signin
required path, if they do we continue as before.

It's not 100% clear to me that we need to do this.  All these requests
will have been preceded by a call to the oauth endpoints to obtain the
access token that is used in this request.  The oauth endpoints check
that the user has signin permission on the application so it should not
be possible to obtain an access token without this permission. However,
I'm not sure how long lived our access tokens are, and by making sure
each api request to signon checks this, we are able to invalidate
requests as soon as the user's permission is revoked rather than waiting
for their access token to expire.