Introduce BulkGrantPermissionSet model and job
We follow the pattern established by BatchInvitation and have a model that
is stored in the DB with a list of permissions, and then we create a job
that performs the work.  The work in this case is to iterate through all
users and grant them each permission in the permission set.  We use a job
for this because there are 1,000s of users and iterating through them all
in a single request/response cylce seems unlikely to be fast.