Add extra info in account access logs
Ip address and browser of user is now recorded on each successful
sign in. 'browser' gem is used to retrieve the browser that the
user is using to sign in.

When a new user account is created by an existing admin, the new
account has a record of who it was invited by.

These extra logging should make it easier to retroactively
trace user behaviour in the case of an incident.