Store ip address and user agent string on logins
The ip address and user agent when user logs into signon is no
longer stored in the trailing message of an event log and
instead the IP address is stored as an integer in event_logs
and the user agent string is stored in a new user_agents tables.
An association is therefore made between the event logs and user agent.
This converts the information from unstructured text into something more
structured.