Model DWP org structure correctly
For: https://trello.com/c/z976OdQw/274-model-dwp-correctly-within-signon

Checks that the children and grandchildren of the "Department for Work and
Pensions" (DWP) have the right ancestry assigned to them.  We also make
sure the children orgs that should be closed are closed.  There were two
orgs that are shared between Cabinet Office (CO) and DWP that we have
decided to assign to the CO and so we update those too.