Model Home Office correctly
For: https://trello.com/c/Rmt2hrLQ/281-model-home-office-correctly-within-signon

Checks that the children and grandchildren of the "Home Office" (HO) have
the right ancestry assigned to them.  We also make sure the children orgs
that should be closed are closed.

There is one closed org in this list, "Chief Fire and Rescure Adviser"
that we expect to be missing, so it's not a problem when running if this
is the case.