Add canary test to warn if we upgrade doorkeeper
We extracted the code in SigninRequiredAuthorizationsController from
version 4.2.0 of doorkeeper.  If we upgrade doorkeeper we want to be
reminded to review the new version of doorkeeper and compare our
controller against the default one to see if we need to make any
changes.  If we are lucky it becomes easier to augment the auth
behaviour without overriding the whole controller.

The intention is that someone does a gem upgrade, this test fails and
reminds them to do the check.  When they're happy they make any relevant
changes to our implementation and tests and then change the
EXPECTED_DOORKEEPER_VERSION constant in
SigninRequiredAuthorizationsController to match the new version and the
test will pass.