Introduce migration to create default permissions
After our [audit of applications that use signon][audit] we found that we
only need two default permissions:

1. signin on the support app
2. signin on the content preview app

Both of these apps are intended to be used by every user so we make them
default permissions, and queue up a bulk grant to make sure all existing
users have these permissions.  There were other apps that did not check
the signin permission, but all of these turned out to be an error of
implementation.

[audit]: https://docs.google.com/spreadsheets/d/1ypHdDHaHJZELkdQC3xQprMxJpkPCLTfnJPsEtPhhENI/edit#gid=0