Include Pundit in signin_required_authorizations_controller
Without this we can't render the signon layout for any of the paths
through the controller where it will cause a render instead of a redirect.
We need it because the layout uses the policy helper that pundit provides
to make sure we don't show non-admin users links to the list of users or
applications or other things they're not allowed to do or see.