Be explicit about index key length on user_agents
MySQL only supports key lengths of 767 bytes (which means about 255
characters as a character can take up to 3 bytes to store).  In MySQL
5.5 it would silently truncate the key for you, however in MySQL 5.6
it won't (see: https://bugs.mysql.com/bug.php?id=68453).  We plan to use
MySQL 5.6 on our new AWS infrastructure and this migration won't load
because of this bug.  We can work around it by explicitly telling MySQL
to only use the first 255 chars of the field rather than assuming it
would do this for us.