Update parameters within controllers
Due to the fact ActionController::Parameters no longer inherits from Hash - all instances of parameters being passed into UserParameterSanitiserneed to be a hash.

To convert these objects into a hash, they are required to have the fields you wish to include permitted. This commit permits all the required fields within the controllers so that values are passed correctly