Change the password controller to check for errors
Due to a change when updating devise (https://github.com/plataformatec/devise/pull/4456/files)
blank passwords no longer get set on the resource when trying to reset
the password with a blank string. This change meant that the call to
resource.valid? now returned true rather than false, as the password was
nil - which is valid.

To fix this, the update method now checks for the presence of errors
rather than checking if the resource is valid.

Additionally, the error regarding password strength is no longer
displayed when submitting a blank password so that test has been removed