Restructure inviting users integration test
The "a normal user being invited by an admin" and "for an end-user by an
admin" contexts were basically the same test (e.g. what can an admin user
do on the invitation page) so we only need one of them.  The only
difference is the assertion in the "a normal user..." one that admins
can't see the Role dropdown, so we merged that into the "for an
end-user..." one as a separate test and removed the "a normal user..."
one.  The "for non-superadmins" context can also be removed by folding it
into the "for an end-user..." context, and the reverse version (that
super admins can see it) is added as a test to the "for an admin by a
superadmin" context.

While adding the reverse version of the 2sv test we spotted that it failed
because the label of the control is not "Require 2-step verification" but
"Ask user to set up 2-step verification", and so the original test had
been passing for the wrong reasons all this time (or at least since the
text on the label changed).  The new test should protect us against that
happening again.

Finally, we rename the remaining contexts to "as an admin" and "as a
superadmin" to make them clearer (the superadmin one still only creates
an admin user, even though it has the role dropdown and can create any
type of user, but it's enough to assert that the role choice is carried
through - we don't really need to create all different types of user to
prove it works).