Fix setting permissions when inviting users
As part of the rails 5 upgrade we had to changee the
`InvitationsController#unsanitised_user_params` method to handle the more
robust paramiter sanitisation in rails 5.  Unfortunately we forgot to add
the `supported_permission_ids` param to the permitted params and so it
was ignored when we try to create the user instance, meaning the user only
gets the default permissions, not the ones selected by the admin.

We had not explicit test for this, so we add some as well as fixing the
issue.