Regenerate schema.rb using modern rails
Modern rails has changed the format of schema.rb so it's useful to
regenerate it as a separate commit/PR so that the next PR with a migration
doesn't incur a large diff-cost simply for adding a column.

This regenerated schema has revealed that we some of our tables still use
latin1 instead of utf8, but we'll fix that separately.