Add organisation_slug to batch_invitation_user
This is to support batch inviting users to separate orgs, rather than
all users in a batch being invitied to one org.  We use slug because
that's what we'll ask people to put in the CSV, and we fallback to
the org of the batch only if the value of the slug is nil or blank.
If it's there but not a real slug then we raise an error on the
assumption that it's better to complain than it is to invite a user
to an org you didn't suspect.