Add integration spec for batch inviting with org info
We tweak the UI slightly to let super admins know that they have this
abiilty.  We also clarify for both super admins and admins what the CSV
file should look like and what the organisation dropdown does which should
make it clearer why an admin could upload a field and import all the users
but a super admin might get errors for some (because admins ignore any
organisation info in the file, but superadmins don't).