Stop running the tests against postgresql on CI
We ran the tests on CI against mysql and postgresql because we deploy
signon against a mysql db in our infrastructure, but other people that
use signon do so against a postgresql db in their infrastructure.  We
wanted to give them confidence that changes we made didn't break their
version of signon, and it gave us some confidence that when we finally
decide to migrate our signon infrastructure to postgresql all would be
well.

However, in the previous commit, we regenerated the rails `db/schema.rb`
and this file is no longer db-agnostic.  This means the postgresql tests
no longer pass because postgresql can't interpret the mysql-specific
options in `db/schema.rb`.  In the interests of not blocking us we've
decided that stopping CI from running against postgresql is the best
option, even if it impacts other teams.  We do plan to migrate signon
to postgresql sooner rather than later, at which point we'll regenerate
the schema from a postgresql instance, and stop running the CI tests
aginst mysql.