Ensure all users have signin permission for Asset Manager
This database migration is heavily based on this one [1] which did the
same thing for the Support and Content Preview apps.

The migration does two things:

1. Makes `signin` a default permission for the Asset Manager app. This
means that new users will automatically be granted this permission.

2. Queues up jobs to grant the permission to all existing users.

We need this because we're using SSO in Asset Manager to protect draft
assets. In the same way that all users are given `signin` permission for
the Content Preview app in order to be able to view draft content, all
users will need `signin` permission in order to be able to view draft
assets.

Ideally we'd create a new Assets Preview app in Signon for this
permission to mirror the Content Preview app. However, until we split
Asset Manager into frontend & backend apps (or at least two instances of
the same app with different configs), we're going to need use the
existing Asset Manager `Doorkeeper::Application`.  This is because the
`gds-sso` gem only supports authenticating against a single Signon app
and the Asset Manager app is already using `gds-sso` to protect its API.

We're OK with this, because the API routes are not publicly accessible,
even via the new `draft-assets` host. However, it might be worth
considering changing Asset Manager so it checks that the authenticated
user for API access matches ones of the Signon API users for one of the
publishing apps.

[1]:
https://github.com/alphagov/signon/blob/c4b928af9ac85069362f7e332f0285fae1086467/db/migrate/20171107153427_add_default_permissions_and_bulk_grant_them_to_all_users.rb