Make sure we default to 'A' when listing users
If no letter is supplied, the new version of alphabetical_paginate will
list users that start with punctuation rather than 'A' if there are any
users whose name starts with punctuation instead of a letter.  It turns
out that we have a handful of users whose name starts with a space
character and so we need to force 'A' as the default letter if it's not
specified as otherwise it looks like we've paginated a weird list.

We also pass a duplicated version of `ALPHABETICAL_PAGINATE_CONFIG` into
the `alpha_paginate` call because the new version will mutate the supplied
hash, and we probably don't want that.