Ensure BatchInvitation gets saved
At the moment we call `save` which returns true or false depending
on whether the save actually worked. In the case when it fails, we
ignore that and then continue to use it as a foreign key of the
BatchInvitationUser model, which then fails because the
BatchInvitation isn't saved, and when that tries to save it fails
again.

This doesn't solve the problem that we're seeing with CSV imports
not working, but it should push the problem further up so we can
identify what's going wrong and fix it.