Allow granting same permission multiple times for invitation
We should check that the batch invitation doesn't already have
said permission before adding it to prevent a Rails validation
error when saving the batch invitation. This copies the same
`grant_permission` and `grant_default_permissions` methods that
already exist for the `User` model.

This solves a problem where the user interface for creating batch
users doesn't show the default permissions which are implicitly
granted to new users. Therefore it was possible to create a batch
user configuration with the same supported permission declared twice
which would fail Rails validation with the unhelpful error of
"The change you wanted was rejected".