Don't show an error when signing in
Currently, a user who visits `signon.publishing.service.gov.uk` will be
confronted with a red flash alert message telling them "You need to
sign in before continuing".

This commit removes the message, because the fact that they've arrived
at a login page should be enough of a hint. The red message makes it
look like something went wrong, but this scenario happens 99.9% of the
time, so there's no need to alarm them.

https://trello.com/c/05wMNlxg