Rename "passphrase" to "password"
"Passphrase" is an unusual name for "password". There's no commit
history that explains the reasoning for calling it "passphrase", but we
think it may have something to do with encouraging users to use a
phrase instead of a word a la https://xkcd.com/936 (which came out in
2011, the
approximate birth year of Signon).

This commit changes "passphrase" to "password". This means it's more
consistent with other government services, which also use the word
"password" (https://design-system.service.gov.uk/patterns/passwords).

https://trello.com/c/05wMNlxg/180-redesign-rebuild-signon