Implement new design for dashboard
This is a quick implementation of a dashboard redesign. The
functionality is mostly the same as before.

I've removed the roles of the user, because I think this doesn't tell
the user anything at the moment. Users that care about the roles of a
user are probably all admins, and will have access to their own
profile, which also lists the roles.