Ensure path exists before creating usage report
When we create the usage report csv, we should ensure the path exists
before creating the csv. The current behaviour will error if the
path does not exist and as such, the report will not be created.

Although the actual fix is only a one-line, refactor the code
to inject File and FileUtils to the constructor of UsagePresenter
so we can more easily test interactions with the file system.

I did explore using FakeFS as well as Minitest::Filesystem, but
decided against using them to avoid introducing extra dependencies.