This fixes some bugs in commit c81f45305e760dfdb79b5cec7ac4d2901d01f5db
The previous commit works in a rails console but not when deploying.
This is due to it not having a proper definition of
Doorkeeper::Application and it having the method
Doorkeeper::Application.supported_permissions. By adding the
definition and using a where command instead, this fixes the issue.

Zendesk: https://govuk.zendesk.com/agent/tickets/3634856