Write an event to audit log for 2i granting Rake task
This rake task was granting users permission to perform 2i in Collections
Publisher but did not log anything in Signon's audit log.  This adds a
log message, also noting that it was performed by Rake task.

An email address must be passed in as a parameter (initiator_email_address).
This user will be the one who is logged as having performed the action.