Re-work permissions granting for 2i users in Collections Publisher
We were previously dealing with two scenarios:
 - User had any permission for the application, so was granted '2i reviewer' only
 - User had no permissions for the application, so was granted 'signin', 'GDS Editor' and '2i reviewer'

This failed users who had the 'GDS Editor' permission, but not 'signin', since they still would not
be able to log into the application (lacking 'signin').

This changes the logic so we grant all the permissions that a relevant user doesn't currently have.
Everyone listed is a GDS Content Designer and is accredited for 2i review, so they should be granted
all three permissions.