Monkey patch Doorkeeper scopes
Doorkeeper 5.2.0rc3 introduces this change
https://github.com/doorkeeper-gem/doorkeeper/pull/1277
which assumes that either the request will include a scope param,
or otherwise uses the servers default scope. Scope validation
will now fail if the scope is blank.

As we are not using scope options, we need to monkey patch
this process to the previous behaviour, which will allow the
scope to be blank without failing validation.