Add print stylesheets to assets pipeline
We have been skipping upgrades of govuk_publishing_components for a long time
because the gem expects consuming apps to have print.css in its asset pipeline.
This adds it to the pipeline, with barebones print support as provided by
the gem, which allows us to upgrade, and thus benefit from all the other
good things the gem has to offer.