Add `with_permission` scope to User
This scope returns all users that have a specific permission. I'm using
UserApplicationPermission here because permissions are specific per user
and per application. Each permission that a specific user has for a
specific application is represented by and instance of
UserApplicationPermission.

This scope will be used in the UI to filter users by the permissions
they have.

Trello card: https://trello.com/c/kuKk7npQ/2051-5-enable-filtering-of-managing-editors-in-signon