Define mail methods on UserMailer to avoid using the default Devise methods
The default mailer methods defined by Devise will not use the correct
methods to interface with Notify: we need to call `view_mail` and pass
`template_id`, and overriding the `mail` method itself at a lower level
doesn't work.

This involves creating one method for each type of mail sent by Devise
(and `devise_invitable`) as well as templates for them — some of these
can simply be moved, because we previously overrode the templates but
not the methods. (However, rather than simply moving the plaintext
template, I've converted the slightly-different HTML template to
plaintext, as Notify will be converting it _back_ into HTML when it's
sent).

This also partially reverts #1482, because that PR [overrode][1] a method
unnecessarily. (For reference: `super` there calls `devise_mailer` and
`devise_mailer` is set to `UserMailer`, so calling `UserMailer`
explicitly is unnecessary.)

[1]: https://github.com/alphagov/signon/pull/1482/files#diff-4676c008b11a5480d73d4a6de01e45b9R160