Remove Jasmine Helper
This was added over 6 years ago:
https://github.com/alphagov/signon/commit/04e34a8eef138e99d92344178d3ebc60c344a005#diff-54dc43b93174691dc226df624651d227R1-R4

We explicitly stopped PhantomJS from being downloaded, to use the
PhantomJS that is already on the PATH. But it's not clear why we
did this (nor is it clear from the PR description in #244).

It's been working fine, but recently the CI build has broken,
whereas the same code works fine in Docker. So the less environmental
manipulation we brittly rely on, the better.

Using the PhantomJS included with Jasmine seems to fix the build,
so we no longer need the helper file.