Add govuk_editor permission to all Publisher users
This adds a task which when run adds the `govuk_editor` permission to
all users who currently have access to the Publisher application. This
is a necessary step towards adding a new permission later which will
only give permission to edit Welsh content in Publisher.

Publisher will be updated to recognise the `govuk_editor` permission as
granting the ability to perform every action on an edition. The upcoming
`welsh_editor` permission will only grant access to editing Welsh
language editions.

Once run the task can be deleted.