Replace phantomjs with Selenium driven Chrome
This commit changes this repo to use the govuk_test gem which provides
a Chrome headless Selenium configuration for Capybara. It then uses
the govuk_test to configure Jasmine for selenium to use the same Chrome
headless configuration.

This has been done to reflect that GOV.UK convention is to use Chrome
Headless over PhantomJS - PhantomJS development has paused. The use of
govuk_test prepares this repo for an upcoming govuk-docker update [1].

[1]: https://github.com/alphagov/govuk-docker/pull/394