Remove need for REDIS_HOST and REDIS_PORT env vars
This is one of the few GOV.UK projects that uses this old approach to
configuring Redis - whereas most of the other ones use the approach
supported natively by the Redis gem [1] of using REDIS_URL. Removing
this will allow us to remove the need for REDIS_HOST and REDIS_PORT env
vars in our architecture.

For the most part this change just simplifies this app.

[1]: https://github.com/redis/redis-rb