Manually set APHAs parent to Defra
This updates the organisation fetcher to manually fix APHA to set its
parent to Defra. This introduces tech debt, but it was deemed to be the
most practical option. Defra is a better parent for APHA than what it
ends up being set to (The Welsh Government) as often editors in Defra
will need access to content published by APHA.

It's worth noting, we've introduced a migration to make this change on
three different occassions, but each time it's wiped when the fetcher
runs:

- https://github.com/alphagov/signon/pull/574
- https://github.com/alphagov/signon/pull/1157
- https://github.com/alphagov/signon/pull/1481

At the moment the chosen parent will be the one that is last in the list
ordered alphabetically. We wondered whether changing it to be the first
in the list as stored in Whitehall was better, but although it fixes the
parent for APHA, it introduces some different parents to other
organisations that doesn't seem quite right. We felt that this wasn't a
decision for us to make.