Remove PostgreSQL configuration
Once upon a time this app was run by GDS with a MySQL database and by
another government department using a PostgreSQL database. Because of
this the app used to support and test against both databases.

Over time this came undone as Rails made db/schema.rb no longer schema
agnostic [1], which led to the removal of PostgreSQL tests and an
expectation that any PostgreSQL users would have to make manual changes
to the code base for usage.

This commit removes the remnants of that world by removing the remaining
code that is related to supporting both types of DB. This app is now
entirely specific to MySQL.

[1]: https://github.com/alphagov/signon/pull/647