Add Authorisations API for api_user tokens
This API will enable SecretsManager Rotation Lambdas [1]
to create, test, and rotate api user bearer tokens for
applications.

This will enable us to automate the creation and rotation
of bearer tokens, a source of toil. This is also an
improvement to our security posture, since we will rotate
credentials more frequently.

Not covered here is the deletion of old, expired secrets.
I expect that Signon will delete expired tokens automatically
with a worker.

A Concourse job will be responsible for cycling tasks in ECS.

This change is not intended to impact the current platform.

This does introduce a pretty dangerous API. This is protected
by a token stored in SecretsManager, and the API endpoints
in ECS are IP restricted to our AWS NAT gateway.
When a bearer token is not set, as in our current platform,
the API will always return a 401 Forbidden error.