Remove explicit `sidekiq` and `redis` dependencies
We need to upgrade the `mail-notify` dependency, as it is
required to bump to Rails 6.1 (see next commit). However,
`mail-notify` 1.0.4 depends on Sidekiq 5.2.9, so we were stuck on
1.0.2 because our version of Sidekiq was 5.2.8, because it
depended on Redis < 4.2, which was locked at 4.2.2.

We don't need the explicit `sidekiq` dependency, as we're already
pulling in `govuk_sidekiq`. And `redis` is a dependency of
`sidekiq`, so also doesn't need to be explicitly declared by us.