Add get OAuth application API endpoint
This adds the endpoint /api/v1/applications?name=Publishing+API
which will return the OAuth ID and Secret for a Doorkeeper App.

The endpoint is IP-restricted and requires a bearer token.

This will be used by the Concourse deploy pipeline when deploying
apps. It's not sufficient to provide only a create endpoint, since
we can be in a state where an application has been created but the
generated secrets don't yet exist in SecretsManager. This endpoint
enables the deploy pipeline to recover from this state.

Trello: https://trello.com/c/elhnwnaf/454-bootstrapping-signon-oauth-app-secrets