Fix Puma setup
Our `jasmine:ci` task calls Puma, which relies on there being a
`tmp/pids` directory. We have a `tmp` directory, but no `pids`
directory, and as `tmp` is git-ignored, it's not straightforward
to ensure that a `pids` subdirectory exists. There doesn't seem
to be a downside to changing where the `pidfile` should live though.