Add API for creating API Users
This will enable the Kubernetes cluster client to create
all required signon resources via an API - previously
we created API Users using a rake task.

The only expected client is a Job running in-cluster. This
new endpoint, like those for applications and authorisations
requires token authentication and is not publicly available.
In our current platform the restriction is implemented
at Nginx [1], in the new platform the firewall is implemented
at the ALB [2].

[1]: https://github.com/alphagov/govuk-puppet/pull/11071/commits/c02e7a86482175fb653613f01653aba74cf33e35
[2]: https://github.com/alphagov/govuk-infrastructure/commit/f0148ce4a505809280d65a219a0887ac0f3fbc90

Trello: https://trello.com/c/fUXJ6X1U/725-create-signon-resources-when-signon-is-deployed