Return 409 Conflict when ApiUser email is already taken
Currently we return a 400 as Devise uses an ActiveRecord validation
rather than a DB constraint to validate uniqueness of a user based
on email [1].

ActiveRecord would only raise a ActiveRecord::RecordNotUnique error
when a DB exception is raised (i.e. a db uniqueness constraint is violated) [2].

Devise using a ActiveRecord validation rather than a DB constraint is
inconvenient since we want to return a 409. We could skip validations
but there are some validations (e.g. on email format) which we want to keep.

The solution I've gone for is to parse the model error and raise the RecordNotUniqe
error in the controller. It's not perfect but it makes the ApiUsers controller
behaviour consistent with the Application controller (Applications provided by
the Doorkeeper library).

[1]: https://github.com/heartcombo/devise/blob/8593801130f2df94a50863b5db535c272b00efe1/lib/devise/models/validatable.rb#L36
[2]: https://api.rubyonrails.org/v6.1.3.1/classes/ActiveRecord/RecordNotUnique.html