Rewrite password-strength-indicator without jQuery
This is the only script on the design system layout of the app that uses
jQuery. We are due to remove jQuery from govuk_publishing_components so
rewriting this allows dropping the jQuery dependency.

The change here modernises the approach to writing GOV.UK to use the
module system and replaces the use of jQuery.

In rewriting this I've removed various interplay between HTML and JS
that was quite complex (and possibly inaccessible). Previous all the
errors were part of the HTML DOM but they were conditionally shown based
on classes. I've changed it to just add the errors via JS.

This also removes a few behaviours that no longer had an affect. There
was an input hidden of the password scoring that there is no server side
code for and there were some indicators of good password behaviour where
there is no longer a UI.