Remove link to make a support request when logged out
This link is rendered to users who are not logged in. However the
support system is behind Signon, so requires a user to be logged in to
make a request. This leads to them being sent back to the same login
page.

Removing the link to remove the continuous loop.