Symlink /app/tmp to /tmp in the Docker image.
This covers the general case for Ruby gems wanting to write to what
should be a read-only filesystem. We've configured most of the cache
locations and whatnot already, but it seems there's no standard setting
for Ruby/Rails to tell it where to write temporary files.

We create the symlink in the build stage as well as the final stage,
because things like `rails assets:precompile` write stuff to `app/tmp`
(which thankfully we don't need) and that would make it a pain to create
the symlink later.