Rewrite redirect and home uris for Applications
This adds the ability to rewrite parts of the redirect and home uris for
Oauth Applications. This allows us to run Signon in our new
infrastructure alongside the existing instance of Signon and redirect
users to the correct domains for applications running in the new
infrastructure. This will remove the need to create seperate Oauth
Applications for instances of applications in the new infrastructure.